Fix: 코드리뷰 반영(패키지 구조 수정)
