+++ backend/src/main/java/project/backend/domain/dm/dmMessage/dto/DmMessageRequest.java
@@ -1,0 +1,11 @@
+package project.backend.domain.dm.dmMessage.dto;
+
+import project.backend.domain.chat.chatmessage.entity.MessageType;
+
+public record DmMessageRequest(
+	String receiverUsername,
+	String content,
+	MessageType type
+) {
+
+}
