--- backend/src/main/java/project/backend/domain/member/dmRoom/dmMessage/dto/DmMessageRequest.java
@@ -1,11 +1,0 @@
-package project.backend.domain.member.dmRoom.dmMessage.dto;
-
-import project.backend.domain.chat.chatmessage.entity.MessageType;
-
-public record DmMessageRequest(
-	String receiverUsername,
-	String content,
-	MessageType type
-) {
-
-}
