--- backend/src/main/java/project/backend/domain/member/dmRoom/dmMessage/listener/DmMessageEventListener.java
@@ -1,27 +1,0 @@
-package project.backend.domain.member.dmRoom.dmMessage.listener;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.messaging.simp.SimpMessagingTemplate;
-import org.springframework.scheduling.annotation.Async;
-import org.springframework.stereotype.Component;
-import org.springframework.transaction.event.TransactionPhase;
-import org.springframework.transaction.event.TransactionalEventListener;
-import project.backend.domain.member.dmRoom.dmMessage.app.DmMessageService;
-import project.backend.domain.member.dmRoom.dmMessage.dto.DmMessageResponse;
-
-@Component
-@RequiredArgsConstructor
-public class DmMessageEventListener {
-
-	private final DmMessageService dmMessageService;
-	private final SimpMessagingTemplate messagingTemplate;
-
-	@Async
-	@TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
-	public void handleDmMessageSent(DmMessageResponse dmEvent) {
-		DmMessageResponse response = dmMessageService.saveEvent(dmEvent);
-
-		messagingTemplate.convertAndSend("/topic/dm/" + response.roomId(), response);
-	}
-
-}
