+++ backend/src/main/java/project/backend/domain/member/friend/app/FriendService.java
@@ -9,18 +9,18 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import project.backend.auth.dto.MemberDetails;
 import project.backend.domain.member.app.MemberService;
-import project.backend.domain.member.dmRoom.app.DmRoomService;
+import project.backend.domain.dm.dmRoom.app.DmRoomService;
 import project.backend.domain.member.entity.Member;
 import project.backend.domain.member.friend.dao.FriendRequestRepository;
 import project.backend.domain.member.friend.dao.FriendsListRepository;
 import project.backend.domain.member.friend.dto.FriendResponse;
 import project.backend.domain.member.friend.entity.FriendRequest;
-import project.backend.domain.member.notification.dto.NotificationDto;
+import project.backend.domain.notification.dto.NotificationDto;
 import project.backend.domain.member.friend.entity.Friends;
-import project.backend.domain.member.notification.app.NotificationService;
+import project.backend.domain.notification.app.NotificationService;
 import project.backend.domain.member.friend.dto.FriendRequestDto;
-import project.backend.domain.member.notification.entity.Notification;
-import project.backend.domain.member.notification.entity.NotificationType;
+import project.backend.domain.notification.entity.Notification;
+import project.backend.domain.notification.entity.NotificationType;
 import project.backend.global.exception.errorcode.FriendErrorCode;
 import project.backend.global.exception.ex.FriendException;
