--- backend/src/main/java/project/backend/domain/member/notification/entity/NotificationType.java
@@ -1,10 +1,0 @@
-package project.backend.domain.member.notification.entity;
-
-public enum NotificationType {
-	FRIEND_REQUESTED,
-	FRIEND_ACCEPTED,
-	FRIEND_REJECTED,
-	WE_ARE_FRIEND_NOW,
-	NEW_DM,
-	CODE_REVIEW
-}
