+++ backend/src/main/java/project/backend/domain/notification/entity/NotificationType.java
@@ -1,0 +1,10 @@
+package project.backend.domain.notification.entity;
+
+public enum NotificationType {
+	FRIEND_REQUESTED,
+	FRIEND_ACCEPTED,
+	FRIEND_REJECTED,
+	WE_ARE_FRIEND_NOW,
+	NEW_DM,
+	CODE_REVIEW
+}
